Bit-independent allocator

+++ lib/alloc.h
@@ -18,7 +18,7 @@
  * start: address of first byte of memory managed by the allocator
  * end: address of the first byte of memory no longer managed by us
  */
-void init_page_allocator(void *allocator, uint32_t start, uint32_t end);
+void init_page_allocator(void *allocator, uintptr_t start, uintptr_t end);
 
 /**
  * Print out all allocations, for debugging.
@@ -35,16 +35,16 @@ void show_pages(void *allocator);
  * return: physical pointer to contiguous pages
  *   NULL if the memory could not be allocated
  */
-uint32_t alloc_pages(void *allocator, uint32_t count, uint32_t align);
+uintptr_t alloc_pages(void *allocator, uintptr_t count, uintptr_t align);
 
 /**
  * Free physical pages.
  * addr: address of range to free
  * count: number of bytes to free
  */
-bool free_pages(void *allocator, uint32_t addr, uint32_t count);
+bool free_pages(void *allocator, uintptr_t addr, uintptr_t count);
 
 /**
  * Mark a memory region as allocated.
  */
-bool mark_alloc(void *allocator, uint32_t start, uint32_t count);
+bool mark_alloc(void *allocator, uintptr_t start, uintptr_t count);
